==> code/01C_plot_NDAR_longitudinal_VABS_MELS.Rmd <==
---
title: "01C plot REVAL subtypes longitudinal analysis (VABS/MELS)"
author: "Veronica Mandelli"
date: "2/8/2022"
output: html_document
---
# REVAL SUBTYPES LONGITUDINAL TRAJECTORIES
in this script we will analyze and visualize longitudinal trajectories of the VABS subtypes discovered in NDA databse.
longitudinal analysis will be carry out for both VABS and MSEL

## load libraries, data, useful functions and set paths
```{r, warning = FALSE, errors = FALSE,message=FALSE}
# set path
work_dir = '/Users/vmandelli/OneDrive - Fondazione Istituto Italiano Tecnologia/vineland_proj_edition'
plot_path = file.path(work_dir,'plots')
data_path = file.path(work_dir,'results')
results_path = file.path(work_dir,'results')
code_path = file.path(work_dir,'code','R')
raw_MSEL_data_path = file.path(work_dir,'data','raw','ndar')

library(wesanderson)
library(ggplot2)
library(openxlsx)

#load useful functions
source(file.path(code_path,"spaghettiPlot_new.R"))
source(file.path(code_path,"spaghettiPlot.R"))
source(file.path(code_path,"spaghettiPlot2.R"))


# load ndar for subject ids
# data
datafile = 'VABS_withcluster_edition_041022.csv'
data <- read.csv(file.path(data_path,datafile))

# renema clusters labels in a more intuitive format * 2= high * 3= med * 1 = low
for (i in (1:length(data$cluster_domain))) {
  if (data[i, 'cluster_domain'] == 1){
    data[i,'cluster'] = "high"
  }
  else if (data[i, 'cluster_domain'] == 0){
    data[i,'cluster'] = "med"
  }
  else if (data[i, 'cluster_domain'] == 2){
    data[i,'cluster'] = "low"
  }
}

# define subject list
subect_list = data$subjectkey
```

# VINELAND SCALES OF ADAPTIVE BEHAVIOR
```{r,warning=FALSE,error=FALSE,message=FALSE}
# load files of longitudinal NDA VABS

file = 'data_NDA_wrangled.csv'
data_long_ = read.csv(file.path(data_path,file))
col2use = c("subjectkey","interview_age","collection_id", "communicationdomain_totalb","livingskillsdomain_totalb",
            "socializationdomain_totalb","motorskillsdomain_totalb")
data_long = data_long_[,col2use]   

# select subjects only if VABS labels are present (those included into the reval stratification model)
data_long_sub = data_long[data_long$subjectkey %in% subect_list,]

# merge cluster labels
data_long_sub_clust = merge(data_long_sub,data[,c("subjectkey","cluster")], by="subjectkey")
data_long_sub_clust = data_long_sub_clust[!duplicated(data_long_sub_clust), ]
sum(table(data_long_sub_clust$subjectkey)>1)

# merge the format
file = 'edition_all_data.csv'
edition_data =  read.csv(file.path(data_path,file))
edition_data =  edition_data[,c("edition","collection_id","subjectkey","interview_age", "format")]
data_long_sub_clust_edition = merge(data_long_sub_clust, edition_data, by.x=c("subjectkey", "collection_id","interview_age"), by.y=c("subjectkey", "collection_id","interview_age"))
data_long_sub_clust_edition = data_long_sub_clust_edition[!duplicated(data_long_sub_clust_edition), ]

# check how many subjects (Remember in Reval there is also VABS 3, that is why some subject here are missed)
print(length(unique(data_long_sub_clust_edition$subjectkey)))
print(paste("number of subject with longitudinal VABS",sum(table(data_long_sub_clust_edition$subjectkey)>1)))

# correct the format
data2stat = data_long_sub_clust_edition
####### change factors levellll##############################3
data2stat$format =  factor(data2stat$format,level = c('interview','survey','ed_3'))
#####################################
var_list = c("communicationdomain_totalb","livingskillsdomain_totalb",
              "socializationdomain_totalb","motorskillsdomain_totalb")
covname2use = c('formatsurvey','formated_3') 


for (var in var_list){
  #load the beta1 matrix
  beta1 = read.csv(file.path(results_path,paste("beta1",var,".csv",sep='')),row.names=1)
  
  
  
  #create the full model
  full_model_test = model.matrix(~0 + format,
                                   data = data2stat)
    
    colnames(full_model_test) = c('formatinterview','formatsurvey',"formated_3")
    col_name = paste("new_",var, sep='')
    data2stat[,col_name] = as.numeric(t(data2stat[,var]) - t(beta1) %*% t(full_model_test[,covname2use]))
    
}

print(length(unique(data2stat$subjectkey)))
colnames(data2stat) = c( "subjectId","dataset","interview_age",
                         "communicationdomain_totalb","livingskillsdomain_totalb",
                         "socializationdomain_totalb","motorskillsdomain_totalb","cluster","edition","format",
                         "new_communicationdomain_totalb", "new_livingskillsdomain_totalb", 
                         "new_socializationdomain_totalb", "new_motorskillsdomain_totalb")
```

## longitudinal analysis on VABS
```{r,warning=FALSE,error=FALSE,message=FALSE}
#initialize some variablel to use
var2use = c("new_communicationdomain_totalb", "new_livingskillsdomain_totalb", 
           "new_socializationdomain_totalb", "new_motorskillsdomain_totalb")

# for statistic
data2stat$cluster <- factor( data2stat$cluster,levels = c("high","med","low"))
data2stat$dataset <- factor( data2stat$dataset) 


plt=list()
title_list= c("VABS_com","VABS_liv","VABS_soc","VABS_mot")
color = wes_palette("IsleofDogs1")
size = 4

# looping on VABS variables
for (i in 1:length(var2use)) {
  col2use =var2use[i]
  #col2use =vabs2use[i]
  title = title_list[i]
  
  # plot 
  spagh =   spaghettiPlot2( # adding the covariates ## format/dataset
    data2use = data2stat,
    x_var = 'interview_age',
    y_var = col2use,
    subgrp_var = "cluster",
    cov_vars=NULL,
    xLabel = 'age',
    yLabel = "standard score",
    modelType = "linear",
    fname2save = NULL,
    plot_dots = FALSE, plot_lines = TRUE,
    dot_alpha = 1/10, line_alpha = 3/10,
    xLimits = c(0, 85), yLimits = c(25, 125),
    lineColor = "cluster", dotColor = "cluster",
    legend_name = "Slope", maxIter = 500,
    plot_title = title
  )
  
  # change plot color 
  cols2use = color
  font_size = 11
  p_com = spagh$p
  p_com = p_com + scale_colour_manual(values = cols2use) + scale_fill_manual(values = cols2use)
  p_com = p_com + theme_bw(base_size = font_size) +
    theme(plot.title = element_text(hjust = 0.5, size=font_size)) +
    guides(colour = FALSE, fill = FALSE)
  ggsave(file.path(plot_path,paste('NDAR_VABS_long',col2use,'.png',sep="")), p_com, width = 4.5 ,height = 4)
  print(p_com)   
  plt [[i]]= p_com
  
  # do lme analysis
  lme2save = anova(spagh$lme_model)
  
  
  # correct p value for multiple comparison
  for (a in 1:3){
    p = lme2save[a,"Pr(>F)"]
    #print('p')
    #print(p)
    p_adj_ = p.adjust(p, method = "fdr", n = size)
    #print('p_adj_')
    #print(p_adj_)
    lme2save[a,'p_adj_temp'] = p_adj_
  }
  
 
  lme_final_2save = lme2save[,c("Sum Sq","Mean Sq","NumDF","DenDF","F value","Pr(>F)","p_adj_temp")]
  colnames(lme_final_2save)= c("Sum Sq","Mean Sq","NumDF","DenDF","F value","p value","p_adj")                      
  
  #write.xlsx(lme_final_2save, file.path(results_path,'NDAR_VABS_LONG',paste('ANOVA_',col2use,".xlsx")))
}

library(patchwork)
long_vabs_nda = plt[[1]] + plt[[2]] + plt[[3]] + plt[[4]]
ggsave(file.path(plot_path,'VABS_long_NDA.pdf'), long_vabs_nda, width = 6 ,height = 4)
  
```

## post hoc analysis on NDAR - VABS 
pairwise comparison between subtype
```{r,warning=FALSE,error=FALSE,message=FALSE}
# initialize some empty saving matrices
ANOVA_post_hoc = as.data.frame(matrix(ncol=24,nrow=1))

colnames_2save = c("col2use","sub_A","sub_B","SumSq_age","MeanSq_age","NumDF_age","DenDF_age","Fvalue_age","p-value_age",
                   "p_adj_age", "SumSq_cluster","MeanSq_cluster","NumDF_cluster","DenDF_cluster",
                   "Fvalue_cluster","p-value_cluster","p_adj_cluster",
                   "SumSq_age:cluster","MeanSq_age:cluster","NumDF_age:cluster","DenDF_age:cluster",
                   "Fvalue_age:cluster","p-value_age:cluster","p_adj_age:cluster")

colnames(ANOVA_post_hoc) = colnames_2save

#set size for p correction (number of hypothesis)
size  = 3 # 3 HP for each subscale 
sub_poss1 = c("high","high","med")
sub_poss2 = c("med","low","low")
#i=0


# loop in the VABS subscale
for (i in (1:4)) {
  col2use = var2use[i]
  print(col2use)
  
  # post hoc done by hand!
  #loop on subtypes
  for(o in 1:3){
    
    sub1 = sub_poss1[o]
    substypeA =  subset(data2stat,data2stat$cluster==sub1)
    #for (d in 1:3){
    
    sub2 = sub_poss2[o]
    substypeB=  subset(data2stat,data2stat$cluster==sub2)
    
    # define vars
    title= title_list[i]
    x_var = 'interview_age'
    y_var = col2use
    subgrp_var = "cluster"
    maxIter = 500
    df2use = rbind(substypeA,substypeB)
    cov_vars = NULL
    if (!is.null(cov_vars)){
    for (icov in 1:length(cov_vars)){
      if (icov==1){
        covs2use = sprintf("+ %s",cov_vars[icov])
      } else {
        covs2use = sprintf("%s + %s", covs2use, cov_vars[icov])
      } # if (icov==1){
    } # for (icov in 1:length(cov_vars)){
  } else {
    covs2use = ""
  } 
    
    # define the formula
    form2use = as.formula(sprintf("%s ~ %s*%s%s + (1|%s) + (%s|subjectId)",
                                  y_var, x_var, subgrp_var, covs2use, "dataset", x_var))
    
    ctrl = lmerControl(optCtrl=list(maxfun=maxIter), # lmerControl(optCtrl=list(maxfun=maxIter)
                       check.conv.singular = .makeCC(action = "ignore",  tol = 1e-4),check.nobs.vs.nRE = "ignore")
    # run the model
    m_allgrps <- eval(substitute(lmer(formula = form2use, data = df2use, na.action = na.omit, control = ctrl)))
    
    # saving
    lme2save = anova(m_allgrps)

    #correct p value for multiple comparison
    for (a in 1:3){
      p = lme2save[a,"Pr(>F)"]
      p_adj_ = p.adjust(p, method = "fdr", n = size)
      lme2save[a,'p_adj_temp'] = p_adj_
    }
    
    
    #save the post-hoc lme complete
    lm = lme2save[,c("Sum Sq","Mean Sq","NumDF","DenDF","F value","Pr(>F)","p_adj_temp")]
    colnames(lm) = c("Sum Sq","Mean Sq","NumDF","DenDF","F value","p-value","p_adj")
    
    write.xlsx(lm, file.path(results_path,"post_hoc",'VABS_NDAR',paste('Post_hoc_ANOVA_lme_spagh2_NDAR_',col2use,"_",sub1,"_",sub2,".xlsx",sep='')))
    
    vect2save = c(col2use,sub1,sub2,as.list(as.vector(t(lm))))
    ANOVA_post_hoc = rbind(ANOVA_post_hoc,vect2save)
    #}
  }}
write.xlsx(ANOVA_post_hoc[2:length(ANOVA_post_hoc$col2use),],file.path(results_path,"Posthoc_NDAR_VABSlong.xlsx"))
```

# MULLEN SCALES OF EARLY LEARNING (MSEL)
```{r,warning=FALSE,error=FALSE,message=FALSE}

# load files
file = 'mullen_72months_long.txt'
data_mels_long = read.csv(file.path(raw_MSEL_data_path,file))

# select cols
col2use = c("subjectkey","interview_age", "scoresumm_vr_age_equiv" , 
            "scoresumm_fm_age_equiv",  "scoresumm_rl_age_equiv",  "scoresumm_el_age_equiv" )
data_mels_long = data_mels_long[,col2use]   
# drop duplicates
data_mels_long= data_mels_long[!duplicated(data_mels_long), ]

# select only subject used to run reval
data_mels_long_sub = data_mels_long[data_mels_long$subjectkey%in%subect_list,]
data_mels_long_sub_clust = merge(data_mels_long_sub,data[,c("subjectkey","cluster","collection_id")], by="subjectkey")

# take only interview_age <68 monhts (MELS standard format)
data_mels_long_sub_clust_age=data_mels_long_sub_clust[data_mels_long_sub_clust$interview_age<=68,]

# how many subjects?
data_mels_long_sub_clust_age = data_mels_long_sub_clust_age[!duplicated(data_mels_long_sub_clust_age), ]
print(length(unique(data_mels_long_sub_clust_age$subjectkey)))
print(paste("number of subject with longitudinal MSEL",sum(table(data_mels_long_sub_clust_age$subjectkey)>1)))


# data2 stats
data2stat = data_mels_long_sub_clust_age
print(length(unique(data2stat$subjectkey)))

# change colnames
colnames(data2stat) = c( "subjectId", "interview_age","scoresumm_vr_age_equiv" ,"scoresumm_fm_age_equiv", "scoresumm_rl_age_equiv", "scoresumm_el_age_equiv" ,"cluster" , "dataset")
```

### longitudinal analysis on MSEL
```{r,warning=FALSE,error=FALSE}

# for statistic
data2stat$cluster <- factor( data2stat$cluster,levels = c("high","med","low"))
data2stat$dataset <- factor( data2stat$dataset) 
title_list= c("MSEL_VR","MSEL_FM","MSEL_RL","MSEL_EL")
var2use = c("scoresumm_vr_age_equiv" , 
            "scoresumm_fm_age_equiv", 
            "scoresumm_rl_age_equiv",  
            "scoresumm_el_age_equiv" )
color = wes_palette("IsleofDogs1")
plt_mels = list()
size =  3

# loop on MSEL subscales
for (i in 1:length(var2use)) {
  col2use =var2use[i]
  title= title_list[i]
  resp = spaghettiPlot2(
    data2use = data2stat,
    x_var = 'interview_age',
    y_var = col2use,
    subgrp_var = "cluster",
    cov_vars=NULL,
    xLabel = 'age',
    yLabel = "age equivalent",
    modelType = "linear",
    fname2save = NULL,
    plot_dots = FALSE, plot_lines = TRUE,
    dot_alpha = 1/10, line_alpha = 1/10,
    xLimits = c(0, 60), yLimits = c(0, 90),
    lineColor = "cluster", dotColor = "cluster",
    legend_name = "Slope", maxIter = 500,
    plot_title = title
  )
  
  
  # change color 
  cols2use = color
  font_size = 11
  p_com = resp$p
  p_com = p_com + scale_colour_manual(values = cols2use) + scale_fill_manual(values = cols2use)
  p_com = p_com + theme_bw(base_size = font_size) +
    theme(plot.title = element_text(hjust = 0.5, size=font_size)) +
    guides(colour = FALSE, fill = FALSE)
  ggsave(file.path(plot_path,paste('NDAR_MELS_long',col2use,'.png',sep="")), p_com, width = 4.5 ,height = 4)
  # save plot in a list
  plt_mels[[i]] = p_com
  print(p_com)
  
  # anova on lme
  lme2save = anova(resp$lme_model)
  
  #correct p value for multiple comparison
  for (a in 1:3){
    p = lme2save[a,"Pr(>F)"]
    #print('p')
    #print(p)
    p_adj_ = p.adjust(p, method = "fdr", n = size)
    #print('p_adj_')
    #print(p_adj_)
    lme2save[a,'p_adj_temp'] = p_adj_
  }
  
  
  lme_final_2save = lme2save[,c("Sum Sq","Mean Sq","NumDF","DenDF","F value","Pr(>F)","p_adj_temp")]
  colnames(lme_final_2save)= c("Sum Sq","Mean Sq","NumDF","DenDF","F value","p value","p_adj")   
  
 
   #write.xlsx(lme_final_2save, file.path(results_path,"NDAR_MSEL_LONG",paste('ANOVA_',col2use,".xlsx")))
}

long_vabs_nda = plt_mels[[1]] + plt_mels[[2]] + plt_mels[[3]] + plt_mels[[4]]
ggsave(file.path(plot_path,'MSEL_long_NDA.pdf'), long_vabs_nda, width = 6 ,height = 4)
```

### post hoc comparison for the lme model NDAR - MSEL
```{r,warning=FALSE,error=FALSE,message=FALSE}
# initialize some empty saving matrices
ANOVA_post_hoc = as.data.frame(matrix(ncol=24,nrow=1))

colnames_2save = c("col2use","sub_A","sub_B","SumSq_age","MeanSq_age","NumDF_age","DenDF_age","Fvalue_age","p-value_age",
                   "p_adj_age", "SumSq_cluster","MeanSq_cluster","NumDF_cluster","DenDF_cluster",
                   "Fvalue_cluster","p-value_cluster","p_adj_cluster",
                   "SumSq_age:cluster","MeanSq_age:cluster","NumDF_age:cluster","DenDF_age:cluster",
                   "Fvalue_age:cluster","p-value_age:cluster","p_adj_age:cluster")

colnames(ANOVA_post_hoc) = colnames_2save

#set size for p correction (number of hypothesis)
size  = 3 # 3 HP for each subscale 
sub_poss1 = c("high","high","med")
sub_poss2 = c("med","low","low")
#i=0

#i=0
for (i in (1:4)) {
  col2use = var2use[i]
  #print(col2use)
  
  # post hoc done by hand!
  for(o in 1:3){
    
    sub1 = sub_poss1[o]
    substypeA =  subset(data2stat,data2stat$cluster==sub1)
    #for (d in 1:3){
    
    sub2 = sub_poss2[o]
    substypeB=  subset(data2stat,data2stat$cluster==sub2)
    # define vars
    
    title= title_list[i]
    x_var = 'interview_age'
    y_var = col2use
    subgrp_var = "cluster"
    maxIter = 500
    df2use = rbind(substypeA,substypeB)
    
    # define the formula
    form2use = as.formula(sprintf("%s ~ %s*%s + (1|dataset) + (%s|subjectId)", #(1|dataset) + 
                                  y_var, x_var, subgrp_var, x_var)) # "dataset",
    
    ctrl = lmerControl(optCtrl=list(maxfun=maxIter), # lmerControl(optCtrl=list(maxfun=maxIter)
                       check.conv.singular = .makeCC(action = "ignore",  tol = 1e-4),check.nobs.vs.nRE = "ignore")
    # run the model
    m_allgrps <- eval(substitute(lmer(formula = form2use, data = df2use, na.action = na.omit, control = ctrl)))
    
    # saving 
    lme2save = anova(m_allgrps)
    
    
    #correct p value for multiple comparison
    for (a in 1:3){
      p = lme2save[a,"Pr(>F)"]
      p_adj_ = p.adjust(p, method = "fdr", n = size)
      lme2save[a,'p_adj_temp'] = p_adj_
    }
    

    #save the post-hoc lme complete
    lm = lme2save[,c("Sum Sq","Mean Sq","NumDF","DenDF","F value","Pr(>F)","p_adj_temp")]
    colnames(lm) = c("Sum Sq","Mean Sq","NumDF","DenDF","F value","p-value","p_adj")
    
    #write csv
    
    ## write excel
    write.xlsx(lm, file.path(results_path,"post_hoc",'MSEL_NDAR',paste('Post_hoc_ANOVA_lme_spagh2_NDAR_',col2use,"_",sub1,"_",sub2,".xlsx",sep='')))
    
    vect2save = c(col2use,sub1,sub2,as.list(as.vector(t(lm))))
    ANOVA_post_hoc = rbind(ANOVA_post_hoc,vect2save)
    
  }
  
}

write.xlsx(ANOVA_post_hoc[2:length(ANOVA_post_hoc$col2use),],file.path(results_path,"Posthoc_NDAR_MSELlong.xlsx"))
```
