==> code/02_UCSD_classification.ipynb <==
# UCSD classification on VABS domain

in this notebook UCSD subject, whose T1<72 months, are classified in respective VABS domain clusters. 

The classifier is trained on NDAR testset.

#import libraries
from strat.create_dataset import dataset, prepare_imputation
#from strat.run_rcv import RCV, relabel
import strat.utils as ut
import logging
import numpy as np
import pandas as pd
import logging
from strat.create_long_ndar import build_long
from sklearn.impute import KNNImputer
from sklearn.neighbors import KNeighborsClassifier
import re
from sklearn.preprocessing import StandardScaler
from umap import UMAP
from strat.ucsd_dataset_mod import  create_new,predict_labels,predict_labels_no_umap #prepare_ucsd,
import pickle as pkl
import os
from reval.best_nclust_cv import FindBestClustCV
from sklearn.cluster import KMeans
from strat.visualization import _scatter_plot
from strat.run_rcv import _build_distmat
from sklearn.impute import KNNImputer

## import data and reformat them 
select only a subset of them, dived input features from labels

main_path = '/Users/vmandelli/OneDrive - Fondazione Istituto Italiano Tecnologia/vineland_proj_edition'
data_path = os.path.join(main_path,"data","tidy","nda")
results_path = os.path.join(main_path,"results")
plot_path = os.path.join(main_path,"plot")

# import data from NDA, already splitted into age bins. 
file = "VABS_withcluster_edition_041022.csv"
new_dict_p1 = pd.read_csv(os.path.join(results_path,file),
                        header=0,
                        low_memory=False,
                        index_col='subjectkey')


# split train and test
x_tr = new_dict_p1[new_dict_p1['TR_TS']=='tr']
x_ts = new_dict_p1[new_dict_p1['TR_TS']=='ts']

# imputation of missing: fit on the train and transform the test
impute = KNNImputer(n_neighbors=5)

col2use = ['communicationdomain_totalb','livingskillsdomain_totalb',
           'socializationdomain_totalb','motorskillsdomain_totalb']


X_tr_prepr = pd.DataFrame(impute.fit_transform(x_tr[col2use]), index = x_tr.index)
X_ts_prepr = pd.DataFrame(impute.transform(x_ts[col2use]), index = x_ts.index)


# Import data from UCSD
#ucsd_data = pd.read_csv(os.path.join(ut.tidy_data_folder,'ucsd','tidy_ucsd_long_asd_onlyT1.csv'),index_col='subjectid')

# Import BETA CORRECTION data from UCSD

ucsd_data = pd.read_csv(os.path.join(ut.tidy_data_folder,'ucsd','UCSD_ASD_T1_4reval_edition_corrected.csv'),index_col='subjectid')
ucsd_data.shape


#define the columns I want as features (VABS domanin columns)

ndar_feat = ['communicationdomain_totalb',
             'livingskillsdomain_totalb',
             'socializationdomain_totalb',
             'motorskillsdomain_totalb']
#oringinal columns
ucsd_feat = ['vine_ComTotal_DomStd',
             'vine_DlyTotal_DomStd',
             'vine_SocTotal_DomStd',
             'vine_MtrTotal_DomStd']
# beta corrected
#ucsd_feat = ['new_vine_ComTotal_DomStd',
#             'new_vine_DlyTotal_DomStd',
#            'new_vine_SocTotal_DomStd',
#            'new_vine_MtrTotal_DomStd']

# select NDA test set as the new training set (to train the classifier), so that we avoid circularities of using data multiple times.
new_TR_data =  X_ts_prepr
new_TR_data.colnames = ndar_feat
new_TR_data.head()

# Prepare it to be in the correct format to be the train set
X_tr = new_TR_data
y_tr = x_ts['cluster_domain']

#check it is ok (it should have subject on the rows and the 4 VABS subscles in the columns (COM-DLS-SOC-MOT))
print(y_tr)
X_tr.head()

# Preapre UCSD to be in the correct format to be insert in the classifies to be labels)
X_ts = ucsd_data[ucsd_data['vine_agemo']<=72]

logging.info(f' Selected subjects at their first encounter done at age included in '
                 f'period P1 (<72). Mean (sd) interview age is {round(X_ts.vine_agemo.describe()["mean"], 2)} '
                 f'({round(X_ts.vine_agemo.describe()["std"], 2)}) '
                 f'-- Min / Max:[{round(X_ts.vine_agemo.describe()["min"], 2)}; '
                 f'{round(X_ts.vine_agemo.describe()["max"], 2)}]')

uscd_used_in_reval = X_ts.copy()
X_ts= X_ts[ucsd_feat]


print(X_ts.shape)


# preapre the DBs to insert in as the algorithm requires

#X_tr, y_tr, X_ts, ucsd = prepare_ucsd(new_TR_data, ucsd_data, period='P1') 
# rember python start to count from 0 , so new_dict_p1['P1'][1] 1 here means ts (= test set), 
# we train the algorithm on the NDAR test set to avoid circularity   -- From ucsd_dataselect only subject whose T1 is in P1 Range
# problem is that reval overfit the training set,

## train the classifier

# train the classifier

# initialize che classifier
classifier = KNeighborsClassifier(n_neighbors = 5)

# predict labels using a self build function strat/ucsd_dataset_mod.predict_labels
label_dict, model_fit, Scaler, Umap = predict_labels(X_tr, X_ts, y_tr, classifier)


# remap the labels on the original UCSD dataframe
uscd_used_in_reval['cluster_domain']= uscd_used_in_reval.index.map(label_dict)

## Merge the long_ucsd DB with the cluster + create fake clusters for other diagnosis

take the database in long format with both ASD and TD (other diagnosis) and all the measures (vineland , mullen, ados)

# import ucsd_long all subject
all_ucsd = pd.read_csv(os.path.join(ut.tidy_data_folder,'ucsd','tidy_ucsd_long_allsubj.csv'),index_col='subjectid')
print(all_ucsd.shape)


# a new_dictionary to add also TD DD LD MN and SYB (as cluster labels) :

# TD (typically development)= PrevDDTyp ,PrevLDDTyp,TD  
# DD (developmental delay) = DD+ GDD
# LD (language delay)
# MD (motor delay) = FMD + MD motor delay 
# syb (syblings) = Typ Sib ASD   

diagnosis_dict={"TD":"TD",
                'PrevLDDTyp':'TD',
                'PrevDDTyp':'TD',
                'DD': "DD",
                'GDD' : 'DD',
                'LD': "LD",
                'FMD': 'MD',
                'MD':'MD',
                'Typ Sib ASD':"syb"}

# copy the DB to map colums on
all_ucsd_clust = all_ucsd.copy()
#create a 'sub_id' col to map the label_dict
all_ucsd_clust['sub_id'] = all_ucsd_clust.index

### to create the "cluster_clolumn"
# 1) map col 'sub_id' on label_dict (output from the classifier for ASD subject whose 1st assessment is 
# before 72 months and have a logitudinal data (at least 2 time points)
# 2) map col 'recentDxJ_dxCode' on diagnosis_dict(to get instead of clusters the diagnosis for TD DD LD MD and syb)

all_ucsd_clust['cluster'] = all_ucsd_clust['sub_id'].map(label_dict).to_frame(name='cluster').combine_first(all_ucsd_clust['recentDxJ_dxCode'].map(diagnosis_dict).to_frame(name='cluster'))

print(all_ucsd_clust.shape)
all_ucsd_clust.head()


# save the new DBs
path2save = ut.out_folder
all_ucsd_clust.to_csv(os.path.join(path2save,'ucsd_long_allSubj_clusters_061022.csv'))
ut.out_folder

# check that I have 1185 ASD subject at T1 <72
all_ucsd_clust_t1= all_ucsd_clust[all_ucsd_clust['time']==1]
all_ucsd_clust_t1['cluster'].isin([1,2,0]).sum()


## run the model! in future studies ...

#prepare the cell to run the classifier on every type if DBS

## with UMAP
Model =  pkl.load(open(os.path.join(ut.out_folder, "fittedmodelDomainONLY_NDARts.sav"), 'rb'))
umap = pkl.load(open(os.path.join(ut.out_folder, 'umapDomainONLY_NDARts.sav'), 'rb'))
scaler = pkl.load(open(os.path.join(ut.out_folder, 'scalerDomainONLY_NDARts.sav'), 'rb'))

## no umap
#Model =  pkl.load(open(os.path.join(ut.out_folder, "NO_UMAP/fittedmodelDomainONLY_NDARts.sav"), 'rb'))
#scaler = pkl.load(open(os.path.join(ut.out_folder, 'NO_UMAP/scalerDomainONLY_NDARts.sav'), 'rb'))


def run_KNN_VABS(X):
    subj = X.index

    New_X_ts = umap.transform(scaler.transform(X))
    pred_labels = Model.predict(New_X_ts)
    
    label_dict = {s: lab for s, lab in zip(subj, pred_labels)}
    
    return label_dict

# to check it works the same as before RESULTS should be equal to label label_dict (use the same X_ts (from UCSD))
RESULTS = run_KNN_VABS(X_ts)


### STOP here
