==> code/05_hybridmodel.Rmd <==
---
title: "05 hybrid model"
author: "Veronica Mandelli"
date: "3/23/2021"
output: html_document
---
# HYBRID MODEL
in this script we are going to build the hybrid model by combining the clusters labels from reval with those from VABSnorm

### load packages,data and define paths
```{r, message=FALSE, warning=FALSE, echo=FALSE}
# load packages
library(easypackages)
library(MetBrewer)
library(emmeans)
library(nlme)
library(wesanderson)
library(openxlsx)
libraries('here','ggplot2','gplots','reshape2','gridExtra')

#define paths
work_dir = '/Users/vmandelli/OneDrive - Fondazione Istituto Italiano Tecnologia/vineland_proj_edition'
code_path = file.path(work_dir,'code','R')
uscd_data_path = file.path(work_dir,'results')
result_path = file.path(work_dir,'results')
plot_path = file.path(work_dir,'plots')
color = wes_palette("IsleofDogs1")

#load some fucntions
source(file.path(code_path,"spaghettiPlot_new.R"))
source(file.path(code_path,"spaghettiPlot.R"))
source(file.path(code_path,"spaghettiPlot2_nodataset.R"))
source(file.path(code_path,"alluvial.R"))

# load data
data <- read.csv(file.path(result_path,'UCSD_VABSnorm_longitudinal.csv'))
subjetc_list = unique(data$subjectId)
td_data <- read.csv(file.path(result_path,'TD_only.csv'))
```


### build hybrid model
combining labels from reval and VABS norms
```{r, message=FALSE, warning=FALSE}
for (sub in (1:length(subjetc_list))){
  subset_ = subset (data,data$subjectId==subjetc_list[sub])

  if (subset_[1,'cluster']== "high" &
      subset_[1,'VABSnorms_cluster'] == "high"){
    data[data$subjectId==subjetc_list[sub],'hybrid_cluster']="Rhigh-Vhigh"
  }
  else if (subset_[1,'cluster']== "high" &
           subset_[1,'VABSnorms_cluster'] == "med"){
    data[data$subjectId==subjetc_list[sub],'hybrid_cluster']="Rhigh-Vmed"
  }
  else if (subset_[1,'cluster']== "high" &
           subset_[1,'VABSnorms_cluster'] == "low"){
    data[data$subjectId==subjetc_list[sub],'hybrid_cluster']="Rhigh-Vlow"
  }
  else if (subset_[1,'cluster']== "med" &
           subset_[1,'VABSnorms_cluster'] == "high"){
    data[data$subjectId==subjetc_list[sub],'hybrid_cluster']="Rmed-Vhigh"
  }
  else if (subset_[1,'cluster']== "med" &
           subset_[1,'VABSnorms_cluster'] == "med"){
    data[data$subjectId==subjetc_list[sub],'hybrid_cluster']="Rmed-Vmed"
  }
  else if (subset_[1,'cluster']== "med" &
           subset_[1,'VABSnorms_cluster'] == "low"){
    data[data$subjectId==subjetc_list[sub],'hybrid_cluster']="Rmed-Vlow"
  }
  else if (subset_[1,'cluster']== "low" &
           subset_[1,'VABSnorms_cluster'] == "high"){
    data[data$subjectId==subjetc_list[sub],'hybrid_cluster']="Rlow-Vhigh"
  }
  else if (subset_[1,'cluster']== "low" &
           subset_[1,'VABSnorms_cluster'] == "med"){
    data[data$subjectId==subjetc_list[sub],'hybrid_cluster']="Rlow-Vmed"
  }
  else if (subset_[1,'cluster']== "low" &
           subset_[1,'VABSnorms_cluster'] == "low"){
    data[data$subjectId==subjetc_list[sub],'hybrid_cluster']="Rlow-Vlow"
  }
}



data$hybrid_cluster = factor(data$hybrid_cluster, levels = c('Rhigh-Vhigh','Rhigh-Vmed','Rhigh-Vlow','Rmed-Vhigh','Rmed-Vmed','Rmed-Vlow','Rlow-Vhigh','Rlow-Vmed','Rlow-Vlow'))

# check the clusters labels and numebrosity 

#print(table(data$hybrid_cluster)/5)
#hybrid clusters numerosity
table2save = table(data$hybrid_cluster)/5
write.csv(table2save,file.path(result_path,'UCSD_hybridcluster_numerosity.csv'))

data$cluster = factor(data$cluster,level = c("high","med","low") )
data$VABSnorms_cluster = factor(data$VABSnorms_cluster,level = c("high","med","low") )
print(table (data$cluster,data$VABSnorms_cluster)/5)
table2save = table (data$cluster,data$VABSnorms_cluster)/5
#write.csv(table2save,file.path(result_path,'UCSD_hybridcluster_fromRorV.csv'))
table_VABSnorm =  table(data$VABSnorms_cluster)/5
#write.csv(table_VABSnorm,file.path(result_path,'UCSD_VABSnorm_numerosity.csv'))

#save data for comapring with older_stratification
write.csv(data,file.path(result_path,'final_data_withallclusters.csv'))
```

### prapare variable sto save results from lme
```{r, message=FALSE, warning=FALSE}
# select only some subtypes for further analysis (exclude those with only few subjects in it)
subtype2use = c("Rhigh-Vhigh","Rhigh-Vmed","Rmed-Vmed","Rmed-Vlow","Rlow-Vlow")

# define varibale to use
var2use = c('mullen_VR_AgeEq','mullen_FM_AgeEq','mullen_RL_AgeEq','mullen_EL_AgeEq')

#define data to use
data2use = data[data$hybrid_cluster %in% subtype2use,]
data2use$hybrid_cluster <- factor( data2use$hybrid_cluster,levels =subtype2use)

# saving AIC in a list 
AIC_hybrid = c()
```

### run lme
```{r, message=FALSE, warning=FALSE,error=FALSE}
color = c("#f0be3d", "#04a3bd", "#da7901", "#247d3f", "#20235b")
title_list= c("MSEL_VR","MSEL_FM","MSEL_RL","MSEL_EL")
plt_mels = list()

# for multiple comparison correction
size = 3

for (i in 1:length(var2use)) {
  col2use =var2use[i]
  title= title_list[i]
  resp = spaghettiPlot2_nodataset(
    data2use = data2use,
    x_var = 'mullen_ageMo',
    y_var = col2use,
    subgrp_var = "hybrid_cluster",
    cov_vars=NULL,
    xLabel = 'age',
    yLabel = "age equivalent",
    modelType = "linear",
    fname2save = NULL,
    plot_dots = FALSE, plot_lines = TRUE,
    dot_alpha = 1/10, line_alpha = 1/10,
    xLimits = c(0, 60), yLimits = c(0, 90),
    lineColor = "hybrid_cluster", dotColor = "hybrid_cluster",
    legend_name = "Slope", maxIter = 500,
    plot_title = title
  )
  
  
  #change color 
  cols2use = color
  font_size = 11
  p_com = resp$p
  p_com = p_com + scale_colour_manual(values = cols2use) + scale_fill_manual(values = cols2use)
  p_com = p_com + theme_bw(base_size = font_size) +
    theme(plot.title = element_text(hjust = 0.5, size=font_size)) +
    guides(colour = FALSE, fill = FALSE)
  
  
  
  
  plt_mels[[i]] = p_com
  print(p_com)
  #save the plot
  ggsave(file.path(plot_path,paste(col2use,'_hybrid.png',sep="")),p_com, width = 4.5 ,height = 4)
  
  # saving
  lme2save = anova(resp$lme_model)
  
  
  #correct p value for multiple comparison
  for (a in 1:3){
    p = lme2save[a,"Pr(>F)"]
    #print('p')
    #print(p)
    p_adj_ = p.adjust(p, method = "fdr", n = size)
   # print('p_adj_')
    #print(p_adj_)
    lme2save[a,'p_adj_temp'] = p_adj_
  }
  
  
  lme_definitive_2save  = lme2save[,c("Sum Sq","Mean Sq","NumDF","DenDF","F value","Pr(>F)","p_adj_temp")]
  colnames(lme_definitive_2save) = c("Sum Sq","Mean Sq","NumDF","DenDF","F value","p-value","p_adj")
 write.xlsx(lme_definitive_2save,file.path(result_path,'UCSD_MSEL_HYBRID',paste('ANOVA_lme_spagh2_UCSD_hybrid_MELS_',col2use,".xlsx")))
  # AIC
  AIC_hybrid[i]= AIC(resp$lme_model)
  print( AIC(resp$lme_model))
}
```

### post hoc comparison for the lme model
```{r, message=FALSE, warning=FALSE, error= FALSE}
# select data
data2stat = data[data$hybrid_cluster=="Rhigh-Vhigh" | data$hybrid_cluster=="Rhigh-Vmed" |data$hybrid_cluster=="Rmed-Vmed"|data$hybrid_cluster=="Rmed-Vlow"|data$hybrid_cluster=="Rlow-Vlow",]
data2stat$hybrid_cluster <- factor( data2stat$hybrid_cluster,levels = c("Rhigh-Vhigh","Rhigh-Vmed","Rmed-Vmed","Rmed-Vlow","Rlow-Vlow"));
sub_list =c("Rhigh-Vhigh","Rhigh-Vmed","Rmed-Vmed","Rmed-Vlow","Rlow-Vlow")
var2use =c("mullen_VR_AgeEq","mullen_FM_AgeEq","mullen_RL_AgeEq","mullen_EL_AgeEq")

# initialize some empty saving matrices
ANOVA_post_hoc = as.data.frame(matrix(ncol=24,nrow=1))

colnames_2save = c("col2use","sub_A","sub_B","SumSq_age","MeanSq_age","NumDF_age","DenDF_age","Fvalue_age","p-value_age",
                   "p_adj_age", "SumSq_cluster","MeanSq_cluster","NumDF_cluster","DenDF_cluster",
                   "Fvalue_cluster","p-value_cluster","p_adj_cluster",
                   "SumSq_age:cluster","MeanSq_age:cluster","NumDF_age:cluster","DenDF_age:cluster",
                   "Fvalue_age:cluster","p-value_age:cluster","p_adj_age:cluster")

colnames(ANOVA_post_hoc) = colnames_2save

# multiple comparison size
size = 3 #number of comparisons for number of VABS scales
sub_poss1 = c("Rhigh-Vhigh","Rhigh-Vhigh","Rhigh-Vhigh","Rhigh-Vhigh","Rhigh-Vmed","Rhigh-Vmed","Rhigh-Vmed","Rmed-Vmed","Rmed-Vmed","Rmed-Vlow")
sub_poss2 = c("Rhigh-Vmed","Rmed-Vmed","Rmed-Vlow","Rlow-Vlow","Rmed-Vmed","Rmed-Vlow","Rlow-Vlow","Rmed-Vlow","Rlow-Vlow","Rlow-Vlow")


for (i in (1:4)) {
  col2use = var2use[i]
  print(col2use)
  
  
  for(o in 1:10){
    
    sub1 = sub_poss1[o]
    substypeA =  subset(data2stat,data2stat$hybrid_cluster==sub1)
    #for (d in 1:3){
    
    sub2 = sub_poss2[o]
    substypeB=  subset(data2stat,data2stat$hybrid_cluster==sub2)
    # define vars
    
    #title= title_list[i]
    x_var = 'mullen_ageMo'
    y_var = col2use
    subgrp_var = "hybrid_cluster"
    maxIter = 500
    df2use = rbind(substypeA,substypeB)
    
    # define the formula
    form2use = as.formula(sprintf("%s ~ %s*%s + (%s|subjectId)", #(1|%s) + 
                                  y_var, x_var, subgrp_var, x_var))# "dataset",
    
    ctrl = lmerControl(optCtrl=list(maxfun=maxIter), # lmerControl(optCtrl=list(maxfun=maxIter)
                       check.conv.singular = .makeCC(action = "ignore",  tol = 1e-4),check.nobs.vs.nRE = "ignore")
    # run the model
    m_allgrps <- eval(substitute(lmer(formula = form2use, data = df2use, na.action = na.omit, control = ctrl)))
    
    # saving 
    
    lme2save = anova(m_allgrps)
    
    #correct p value for multiple comparison
    for (a in 1:3){
      p = lme2save[a,"Pr(>F)"]
      p_adj_ = p.adjust(p, method = "fdr", n = size)
      lme2save[a,'p_adj_temp'] = p_adj_
    }
   
    #save the post-hoc lme complete
    lm = lme2save[,c("Sum Sq","Mean Sq","NumDF","DenDF","F value","Pr(>F)","p_adj_temp")]
    colnames(lm) = c("Sum Sq","Mean Sq","NumDF","DenDF","F value","p-value","p_adj")
  
    write.xlsx(lm,file.path(result_path,"post_hoc",'MELS_UCSD_hybrid',paste('Post_hoc_ANOVA_lme_spagh2_UCSD_hybrid_',col2use,"_",sub1,"_",sub2,".xlsx",sep='')))
    #save in a unique table
    vect2save = c(col2use,sub1,sub2,as.list(as.vector(t(lm))))
    ANOVA_post_hoc = rbind(ANOVA_post_hoc,vect2save) 
  }
}



write.xlsx(ANOVA_post_hoc[2:length(ANOVA_post_hoc$col2use),],file.path(result_path,"Posthoc_UCSD_hybrid.xlsx"))
```


### AIC plot
```{r, message=FALSE, warning=FALSE}
# load REVAL AIC matrix from reval and VABSnorm
AIC_matrix = read.csv(file.path(result_path,'AIC_matrix_REVAL_VABSnorm.csv'))
rownames(AIC_matrix)=AIC_matrix$X
AIC_matrix=AIC_matrix[,2:5]

# reformat AIC_hybrid 
AIC_hybrid_df = t(as.data.frame(AIC_hybrid))
colnames(AIC_hybrid_df)=colnames(AIC_matrix)
AIC2plot= rbind(AIC_matrix,AIC_hybrid_df)

# with base graphics
library(reshape2)
AIC2plot$model=c("VABSnorm","REVAL","Hybrid Model")
melt_AIC2plot<-melt(AIC2plot)

# delta AIC
for (i  in 1:4){
  AIC2plot['deltaaic',i] = AIC2plot["REVAL",i]-AIC2plot["AIC_hybrid",i]}

# save the table with only REVAL VABSnorms and HYBRID
AIC2save = round(rbind(AIC_matrix[c('REVAL' ,'VABSnorm'),],AIC_hybrid_df))
write.csv(AIC2save,file.path(result_path,'AIC_matrix_all.csv'))

# plot
ylim_list = list(c(11700,12000),c(10500,10800),c(12500,12800),c(12200,12500))
plot_aic= list()
melt_AIC2plot$model = factor(melt_AIC2plot$model, levels = c('REVAL','VABSnorm','Hybrid Model'))

for (i in 1:4){
  tlim = ylim_list[[i]]
  data2plot = melt_AIC2plot[melt_AIC2plot$variable == var2use[i],]
  plt = ggplot(data2plot,aes(model,value,fill=model))+
    geom_bar(stat="identity",position="dodge")+
    coord_cartesian(ylim = tlim) + 
    scale_fill_manual(values=c("#cccccc","#666666","#000000"))+
    theme_bw()+
    theme(legend.position="none") +labs(x = title_list[i], y = " ")
  plot_aic[[i]]=plt
  print(plt)
}



aic_plot= grid.arrange(
  plot_aic[[1]],
  plot_aic[[2]],
  plot_aic[[3]],
  plot_aic[[4]],
  
  ncol = 4
)
show(aic_plot)    
ggsave(file.path(plot_path,'AIC_plot.pdf'), aic_plot, width = 6 ,height = 4)
```

### get out betas from LME 
betas (of interaction terms AGE*SUBTYPE) represent the rate of growth of the subtype
```{r, message=FALSE, warning=FALSE}
# prepare TD data
colnames(td_data)[which(names(td_data) == "subjectid")] <- "subjectId"
td_data$hybrid_cluster="TD"
subtype_list =c("Rhigh-Vhigh","Rhigh-Vmed","Rmed-Vmed","Rmed-Vlow","Rlow-Vlow")
col2merge =  c("subjectId","mullen_ageMo","mullen_VR_AgeEq","mullen_FM_AgeEq","mullen_RL_AgeEq","mullen_EL_AgeEq","hybrid_cluster")

# prepare data to be on the plot
data2use = rbind(data[data$hybrid_cluster %in% subtype_list,col2merge],td_data[,col2merge])
data2use$hybrid_cluster <- factor( data2use$hybrid_cluster,levels =c(subtype_list,"TD"));


# define my subtype list
subtype_list =c("Rhigh-Vhigh","Rhigh-Vmed","Rmed-Vmed","Rmed-Vlow","Rlow-Vlow","TD") 
#data2use = data[data$hybrid_cluster %in% subtype_list,]
#initialize an empty matrix to save betas in a comprhensible way
betas_matrix= as.data.frame(matrix(nrow=4,ncol=6))
rownames(betas_matrix)= var2use
colnames(betas_matrix)= subtype_list


# run the nestedfor-loop 
for (k in 1:length(subtype_list)){
  #extracting the betas:
  subtype2use = subtype_list[k]
  subset_group = data2use[data2use$hybrid_cluster== subtype2use ,]
  for (i in 1:length(var2use)){
    col2use =var2use[i]
    x_var = 'mullen_ageMo'
    y_var = col2use
    subgrp_var = "hybrid_cluster"
    maxIter = 500
    df2use = subset_group
    
    # define the formula
    # define the formula
    form2use = as.formula(sprintf("%s ~ %s + (%s|subjectId)", #(1|%s) +   # fixed effect --> delate the subtype because I                                                                                         want betas only from one and not a comparison between them?
                                  y_var, x_var, x_var))# "dataset",
    
    ctrl = lmerControl(optCtrl=list(maxfun=maxIter), # lmerControl(optCtrl=list(maxfun=maxIter)
                       check.conv.singular = .makeCC(action = "ignore",  tol = 1e-4),check.nobs.vs.nRE = "ignore")
    # run the model
    m_allgrps <- eval(substitute(lmer(formula = form2use, data = df2use, na.action = na.omit, control = ctrl)))
    
    coef = coef(summary(m_allgrps))[2,1]
    betas_matrix[i,k] = coef
  }
  
}
print(betas_matrix)
write.csv(round(betas_matrix,digits = 2),file.path(result_path,'betas_matrix_hybrid.csv'))
```

###  plot single group trajectories (uncomment here if you want them!)
these plots are going to be insert in the shinyapp. They represent the developmental trajectory of a single subtypes comapred to the TD trajectory
```{r, message=FALSE, warning=FALSE, echo = FALSE}
# # intitalize useful variables
title_list = c("Visual Reception","Fine Motor",
               "Receptive Language","Expressive Language")
var2use = c('mullen_VR_AgeEq','mullen_FM_AgeEq','mullen_RL_AgeEq','mullen_EL_AgeEq')

# plot
for (subtype2use in subtype_list){
  subset_group = data2use[data2use$hybrid_cluster== subtype2use |data2use$hybrid_cluster=="TD" ,]
  subset_group$hybrid_cluster <- factor( subset_group$hybrid_cluster,levels = c(subtype2use,"TD"));
  plt_sub=list()
  i=1
  print(subtype2use)
  for(col2use in var2use){
    #col2use =var2use[i]
    #plot
    print(col2use)
    print(i)
    spagh = spaghettiPlot(df=subset_group,
                          x_var='mullen_ageMo',
                          y_var=col2use,
                          subgrp_var='hybrid_cluster',
                          xLabel = 'chronological age (months)',
                          yLabel = 'age equivalent (months)',
                          modelType = "linear",
                          fname2save = NULL, plot_dots = FALSE, plot_lines = TRUE,
                          ci_band = TRUE, pi_band = FALSE, dot_alpha = 1/10, line_alpha = 1/50,
                          band_alpha = 1/2, xLimits = c(0, 60), yLimits = c(0,90), standardize = FALSE,
                          color = c("#FF9933","#3399CC","#009900","#CC0099"))
    #print(spagh$p)
    spagh$p=spagh$p +theme(legend.position = "none") + labs(title = title_list[i],color = "")

    plt_sub[[col2use]]= spagh$p
    #print(spagh$p)
    i=i+1}
  library(patchwork)
  pfd2save = plt_sub[[1]] + plt_sub[[2]] + plt_sub[[3]] + plt_sub[[4]]
  
  print(pfd2save)
  ggsave(file.path(plot_path, 'shinyplot',paste("MULLEN_",subtype2use,".pdf",sep='')),pfd2save)
}

```
